+++ interface/ui/mainWindow.py
@@ -9,18 +9,18 @@ from PyQt5.QtWidgets import (
     QListWidget,
     QStackedWidget,
 )
-from interface.ui.employees_window import EmployeesWindow
-from interface.ui.reports_window import ReportsWindow
-from interface.ui.tasks_window import TasksWindow
+from interface.ui.employeesWindow import EmployeesWindow
+from interface.ui.reportsWindow import ReportsWindow
+from interface.ui.tasksWindow import TasksWindow
 from models.user import User
 
 
 class MainWindow(QMainWindow):
     def __init__(self, user: Optional[User] = None):
         super().__init__()
-        self.user: Optional[User] = user
         self.setWindowTitle("Management Employees And Tasks Software")
         self.resize(800, 600)
+        self.user: Optional[User] = user
         self.initUi()
 
     def initUi(self):
@@ -39,7 +39,7 @@ class MainWindow(QMainWindow):
 
     def mainPages(self):
         self.pages = QStackedWidget()
-        self.home_widget = QLabel(f"Welcome In System : {self.user.getFullName}")
+        self.home_widget = QLabel(f"Welcome In System : {self.user.getFullName}")  # type: ignore
         self.pages.addWidget(self.home_widget)
         self.employees_widget = EmployeesWindow()
         self.pages.addWidget(self.employees_widget)