+++ main.py
@@ -1,5 +1,5 @@
-from interface.cli.app import start as cli
-from interface.ui.login_page import start as gui
+# from interface.cli.app import start as cli
+from interface.ui.loginPage import start as gui
 
 
 def main():